fix button/ icons click
